requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.12
Flask-Script==2.0.6
py-ms==2.0.0
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0